duck
added duck
also installed probuilder and added a bunch of random shit

As a memo, game puzzle is where you need to blow up the rubbles in the entrance to get out, while not getting blown up yourself

As a bonus: storywise, youre taking control of an unconscious worker and have to get a very important object out of the place. Two endings: You get the object out but kill the worker in the blast, and 2. You get both the object and worker out